Move CTI and CommandManager clients to base client/ directory
Add docstrings to new classes

Remove unused functions
