+++ plugins/content-manager/src/main/java/com/wazuh/contentmanager/client/CTIClient.java
@@ -14,7 +14,7 @@
  * You should have received a copy of the GNU Affero General Public License
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
-package com.wazuh.contentmanager.client.cti;
+package com.wazuh.contentmanager.client;
 
 import org.apache.hc.client5.http.async.methods.SimpleHttpResponse;
 import org.apache.hc.core5.http.Header;
@@ -22,20 +22,33 @@ import org.apache.hc.core5.http.Header;
 import java.net.URI;
 import java.util.Map;
 
-import com.wazuh.contentmanager.client.HttpClient;
 import com.wazuh.contentmanager.settings.PluginSettings;
 
+/**
+ * CTIClient is a singleton class responsible for interacting with the CTI (Cyber Threat
+ * Intelligence) API. It extends HttpClient to handle HTTP requests.
+ */
 public class CTIClient extends HttpClient {
     private static CTIClient instance;
 
     private static final String apiUrl =
             PluginSettings.getInstance().getCtiBaseUrl()
                     + "/catalog/contexts/vd_1.0.0/consumers/vd_4.8.0";
+    private static final String CONTENT_CHANGES_ENDPOINT = "/changes";
 
+    /**
+     * Private constructor to enforce singleton pattern. Initializes the HTTP client with the CTI API
+     * base URL.
+     */
     protected CTIClient() {
         super(URI.create(apiUrl));
     }
 
+    /**
+     * Retrieves the singleton instance of CTIClient. Ensures thread-safe lazy initialization.
+     *
+     * @return The singleton instance of CTIClient.
+     */
     public static CTIClient getInstance() {
         if (instance == null) {
             synchronized (CTIClient.class) {
@@ -47,11 +60,21 @@ public class CTIClient extends HttpClient {
         return instance;
     }
 
+    /**
+     * Fetches content changes from the CTI API.
+     *
+     * @param queryParameters A map containing query parameters to filter the request.
+     * @return A SimpleHttpResponse containing the response from the API.
+     */
     public SimpleHttpResponse getChanges(Map<String, String> queryParameters) {
-        String endpoint = "/changes";
-        return sendRequest("GET", endpoint, null, queryParameters, (Header) null);
+        return sendRequest("GET", CONTENT_CHANGES_ENDPOINT, null, queryParameters, (Header) null);
     }
 
+    /**
+     * Fetches the entire CTI catalog from the API.
+     *
+     * @return A SimpleHttpResponse containing the response from the API.
+     */
     public SimpleHttpResponse getCatalog() {
         return sendRequest("GET", null, null, null, (Header) null);
     }
